prepare sys_limits, and intro page type, add few docs
